changed: put people videodatabase methods in separate file
