+++ xbmc/video/database/VideoDatabase.cpp
@@ -626,64 +626,6 @@ int CVideoDatabase::AddUniqueIDs(int mediaId, const char *mediaType, const CVide
   return uniqueid;
 }
 
-int CVideoDatabase::AddActor(const std::string& name, const std::string& thumbURLs, const std::string &thumb)
-{
-  try
-  {
-    if (NULL == m_pDB.get()) return -1;
-    if (NULL == m_pDS.get()) return -1;
-    int idActor = -1;
-
-    // ATTENTION: the trimming of actor names should really not be done here but after the scraping / NFO-parsing
-    std::string trimmedName = name.c_str();
-    StringUtils::Trim(trimmedName);
-
-    std::string strSQL=PrepareSQL("select actor_id from actor where name like '%s'", trimmedName.substr(0, 255).c_str());
-    m_pDS->query(strSQL);
-    if (m_pDS->num_rows() == 0)
-    {
-      m_pDS->close();
-      // doesnt exists, add it
-      strSQL=PrepareSQL("insert into actor (actor_id, name, art_urls) values(NULL, '%s', '%s')", trimmedName.substr(0,255).c_str(), thumbURLs.c_str());
-      m_pDS->exec(strSQL);
-      idActor = (int)m_pDS->lastinsertid();
-    }
-    else
-    {
-      idActor = m_pDS->fv(0).get_asInt();
-      m_pDS->close();
-      // update the thumb url's
-      if (!thumbURLs.empty())
-      {
-        strSQL=PrepareSQL("update actor set art_urls = '%s' where actor_id = %i", thumbURLs.c_str(), idActor);
-        m_pDS->exec(strSQL);
-      }
-    }
-    // add artwork
-    if (!thumb.empty())
-      SetArtForItem(idActor, "actor", "thumb", thumb);
-    return idActor;
-  }
-  catch (...)
-  {
-    CLog::Log(LOGERROR, "%s (%s) failed", __FUNCTION__, name.c_str() );
-  }
-  return -1;
-}
-
-
-
-void CVideoDatabase::AddLinkToActor(int mediaId, const char *mediaType, int actorId, const std::string &role, int order)
-{
-  std::string sql=PrepareSQL("SELECT 1 FROM actor_link WHERE actor_id=%i AND media_id=%i AND media_type='%s'", actorId, mediaId, mediaType);
-
-  if (GetSingleValue(sql).empty())
-  { // doesnt exists, add it
-    sql = PrepareSQL("INSERT INTO actor_link (actor_id, media_id, media_type, role, cast_order) VALUES(%i,%i,'%s','%s',%i)", actorId, mediaId, mediaType, role.c_str(), order);
-    ExecuteQuery(sql);
-  }
-}
-
 void CVideoDatabase::AddToLinkTable(int mediaId, const std::string& mediaType, const std::string& table, int valueId, const char *foreignKey)
 {
   const char *key = foreignKey ? foreignKey : table.c_str();
@@ -723,41 +665,6 @@ void CVideoDatabase::UpdateLinksToItem(int mediaId, const std::string& mediaType
   m_pDS->exec(sql);
 
   AddLinksToItem(mediaId, mediaType, field, values);
-}
-
-void CVideoDatabase::AddActorLinksToItem(int mediaId, const std::string& mediaType, const std::string& field, const std::vector<std::string>& values)
-{
-  for (const auto &i : values)
-  {
-    if (!i.empty())
-    {
-      int idValue = AddActor(i, "");
-      if (idValue > -1)
-        AddToLinkTable(mediaId, mediaType, field, idValue, "actor");
-    }
-  }
-}
-
-void CVideoDatabase::UpdateActorLinksToItem(int mediaId, const std::string& mediaType, const std::string& field, const std::vector<std::string>& values)
-{
-  std::string sql = PrepareSQL("DELETE FROM %s_link WHERE media_id=%i AND media_type='%s'", field.c_str(), mediaId, mediaType.c_str());
-  m_pDS->exec(sql);
-
-  AddActorLinksToItem(mediaId, mediaType, field, values);
-}
-
-//****Actors****
-void CVideoDatabase::AddCast(int mediaId, const char *mediaType, const std::vector< SActorInfo > &cast)
-{
-  if (cast.empty())
-    return;
-
-  int order = std::max_element(cast.begin(), cast.end())->order;
-  for (const auto &i : cast)
-  {
-    int idActor = AddActor(i.strName, i.thumbUrl.m_xml, i.thumb);
-    AddLinkToActor(mediaId, mediaType, idActor, i.strRole, i.order >= 0 ? i.order : ++order);
-  }
 }
 
 //********************************************************************************************************************************
@@ -1018,57 +925,6 @@ CVideoInfoTag CVideoDatabase::GetDetailsByTypeAndId(VIDEODB_CONTENT_TYPE type, i
   }
 
   return details;
-}
-
-void CVideoDatabase::GetCast(int media_id, const std::string &media_type, std::vector<SActorInfo> &cast)
-{
-  try
-  {
-    if (!m_pDB.get()) return;
-    if (!m_pDS2.get()) return;
-
-    std::string sql = PrepareSQL("SELECT actor.name,"
-                                 "  actor_link.role,"
-                                 "  actor_link.cast_order,"
-                                 "  actor.art_urls,"
-                                 "  art.url "
-                                 "FROM actor_link"
-                                 "  JOIN actor ON"
-                                 "    actor_link.actor_id=actor.actor_id"
-                                 "  LEFT JOIN art ON"
-                                 "    art.media_id=actor.actor_id AND art.media_type='actor' AND art.type='thumb' "
-                                 "WHERE actor_link.media_id=%i AND actor_link.media_type='%s'"
-                                 "ORDER BY actor_link.cast_order", media_id, media_type.c_str());
-    m_pDS2->query(sql);
-    while (!m_pDS2->eof())
-    {
-      SActorInfo info;
-      info.strName = m_pDS2->fv(0).get_asString();
-      bool found = false;
-      for (const auto &i : cast)
-      {
-        if (i.strName == info.strName)
-        {
-          found = true;
-          break;
-        }
-      }
-      if (!found)
-      {
-        info.strRole = m_pDS2->fv(1).get_asString();
-        info.order = m_pDS2->fv(2).get_asInt();
-        info.thumbUrl.ParseString(m_pDS2->fv(3).get_asString());
-        info.thumb = m_pDS2->fv(4).get_asString();
-        cast.emplace_back(std::move(info));
-      }
-      m_pDS2->next();
-    }
-    m_pDS2->close();
-  }
-  catch (...)
-  {
-    CLog::Log(LOGERROR, "%s(%i,%s) failed", __FUNCTION__, media_id, media_type.c_str());
-  }
 }
 
 void CVideoDatabase::GetUniqueIDs(int media_id, const std::string &media_type, CVideoInfoTag& details)
@@ -1451,275 +1307,6 @@ bool CVideoDatabase::GetNavCommon(const std::string& strBaseDir, CFileItemList& 
   return false;
 }
 
-bool CVideoDatabase::GetWritersNav(const std::string& strBaseDir, CFileItemList& items, int idContent /* = -1 */, const Filter &filter /* = Filter() */, bool countOnly /* = false */)
-{
-  return GetPeopleNav(strBaseDir, items, "writer", idContent, filter, countOnly);
-}
-
-bool CVideoDatabase::GetDirectorsNav(const std::string& strBaseDir, CFileItemList& items, int idContent /* = -1 */, const Filter &filter /* = Filter() */, bool countOnly /* = false */)
-{
-  return GetPeopleNav(strBaseDir, items, "director", idContent, filter, countOnly);
-}
-
-bool CVideoDatabase::GetActorsNav(const std::string& strBaseDir, CFileItemList& items, int idContent /* = -1 */, const Filter &filter /* = Filter() */, bool countOnly /* = false */)
-{
-  if (GetPeopleNav(strBaseDir, items, "actor", idContent, filter, countOnly))
-  { // set thumbs - ideally this should be in the normal thumb setting routines
-    for (int i = 0; i < items.Size() && !countOnly; i++)
-    {
-      CFileItemPtr pItem = items[i];
-      if (idContent == VIDEODB_CONTENT_MUSICVIDEOS)
-        pItem->SetIconImage("DefaultArtist.png");
-      else
-        pItem->SetIconImage("DefaultActor.png");
-    }
-    return true;
-  }
-  return false;
-}
-
-bool CVideoDatabase::GetPeopleNav(const std::string& strBaseDir, CFileItemList& items, const char *type, int idContent /* = -1 */, const Filter &filter /* = Filter() */, bool countOnly /* = false */)
-{
-  if (NULL == m_pDB.get()) return false;
-  if (NULL == m_pDS.get()) return false;
-
-  try
-  {
-    //! @todo This routine (and probably others at this same level) use playcount as a reference to filter on at a later
-    //!       point.  This means that we *MUST* filter these levels as you'll get double ups.  Ideally we'd allow playcount
-    //!       to filter through as we normally do for tvshows to save this happening.
-    //!       Also, we apply this same filtering logic to the locked or unlocked paths to prevent these from showing.
-    //!       Whether or not this should happen is a tricky one - it complicates all the high level categories (everything
-    //!       above titles).
-
-    // General routine that the other actor/director/writer routines call
-
-    // get primary genres for movies
-    std::string strSQL;
-    Filter extFilter = filter;
-    if (m_profileManager.GetMasterProfile().getLockMode() != LOCK_MODE_EVERYONE && !g_passwordManager.bMasterUser)
-    {
-      std::string view, view_id, media_type, extraField, group;
-      if (idContent == VIDEODB_CONTENT_MOVIES)
-      {
-        view       = MediaTypeMovie;
-        view_id    = "idMovie";
-        media_type = MediaTypeMovie;
-        extraField = "files.playCount";
-      }
-      else if (idContent == VIDEODB_CONTENT_TVSHOWS)
-      {
-        view       = MediaTypeEpisode;
-        view_id    = "idShow";
-        media_type = MediaTypeTvShow;
-        extraField = "count(DISTINCT idShow)";
-        group = "actor.actor_id";
-      }
-      else if (idContent == VIDEODB_CONTENT_EPISODES)
-      {
-        view       = MediaTypeEpisode;
-        view_id    = "idEpisode";
-        media_type = MediaTypeEpisode;
-        extraField = "files.playCount";
-      }
-      else if (idContent == VIDEODB_CONTENT_MUSICVIDEOS)
-      {
-        view       = MediaTypeMusicVideo;
-        view_id    = "idMVideo";
-        media_type = MediaTypeMusicVideo;
-        extraField = "files.playCount";
-      }
-      else
-        return false;
-
-      strSQL = "SELECT %s FROM actor ";
-      extFilter.fields = "actor.actor_id, actor.name, actor.art_urls, path.strPath";
-      extFilter.AppendField(extraField);
-      extFilter.AppendJoin(PrepareSQL("JOIN %s_link ON actor.actor_id = %s_link.actor_id", type, type));
-      extFilter.AppendJoin(PrepareSQL("JOIN %s_view ON %s_link.media_id = %s_view.%s AND %s_link.media_type='%s'", view.c_str(), type, view.c_str(), view_id.c_str(), type, media_type.c_str()));
-      extFilter.AppendJoin(PrepareSQL("JOIN files ON files.idFile = %s_view.idFile", view.c_str()));
-      extFilter.AppendJoin("JOIN path ON path.idPath = files.idPath");
-      extFilter.AppendGroup(group);
-    }
-    else
-    {
-      std::string view, view_id, media_type, extraField, extraJoin;
-      if (idContent == VIDEODB_CONTENT_MOVIES)
-      {
-        view       = MediaTypeMovie;
-        view_id    = "idMovie";
-        media_type = MediaTypeMovie;
-        extraField = "count(1), count(files.playCount)";
-        extraJoin  = PrepareSQL(" JOIN files ON files.idFile=%s_view.idFile", view.c_str());
-      }
-      else if (idContent == VIDEODB_CONTENT_TVSHOWS)
-      {
-        view       = MediaTypeTvShow;
-        view_id    = "idShow";
-        media_type = MediaTypeTvShow;
-        extraField = "count(idShow)";
-      }
-      else if (idContent == VIDEODB_CONTENT_EPISODES)
-      {
-        view       = MediaTypeEpisode;
-        view_id    = "idEpisode";
-        media_type = MediaTypeEpisode;
-        extraField = "count(1), count(files.playCount)";
-        extraJoin  = PrepareSQL("JOIN files ON files.idFile = %s_view.idFile", view.c_str());
-      }
-      else if (idContent == VIDEODB_CONTENT_MUSICVIDEOS)
-      {
-        view       = MediaTypeMusicVideo;
-        view_id    = "idMVideo";
-        media_type = MediaTypeMusicVideo;
-        extraField = "count(1), count(files.playCount)";
-        extraJoin  = PrepareSQL("JOIN files ON files.idFile = %s_view.idFile", view.c_str());
-      }
-      else
-        return false;
-
-      strSQL ="SELECT %s FROM actor ";
-      extFilter.fields = "actor.actor_id, actor.name, actor.art_urls";
-      extFilter.AppendField(extraField);
-      extFilter.AppendJoin(PrepareSQL("JOIN %s_link on actor.actor_id = %s_link.actor_id", type, type));
-      extFilter.AppendJoin(PrepareSQL("JOIN %s_view on %s_link.media_id = %s_view.%s AND %s_link.media_type='%s'", view.c_str(), type, view.c_str(), view_id.c_str(), type, media_type.c_str()));
-      extFilter.AppendJoin(extraJoin);
-      extFilter.AppendGroup("actor.actor_id");
-    }
-
-    if (countOnly)
-    {
-      extFilter.fields = "COUNT(1)";
-      extFilter.group.clear();
-      extFilter.order.clear();
-    }
-    strSQL = StringUtils::Format(strSQL.c_str(), !extFilter.fields.empty() ? extFilter.fields.c_str() : "*");
-
-    CVideoDbUrl videoUrl;
-    if (!BuildSQL(strBaseDir, strSQL, extFilter, strSQL, videoUrl))
-      return false;
-
-    // run query
-    unsigned int time = XbmcThreads::SystemClockMillis();
-    if (!m_pDS->query(strSQL)) return false;
-    CLog::Log(LOGDEBUG, LOGDATABASE, "%s -  query took %i ms",
-              __FUNCTION__, XbmcThreads::SystemClockMillis() - time); time = XbmcThreads::SystemClockMillis();
-    int iRowsFound = m_pDS->num_rows();
-    if (iRowsFound == 0)
-    {
-      m_pDS->close();
-      return true;
-    }
-
-    if (countOnly)
-    {
-      CFileItemPtr pItem(new CFileItem());
-      pItem->SetProperty("total", iRowsFound == 1 ? m_pDS->fv(0).get_asInt() : iRowsFound);
-      items.Add(pItem);
-
-      m_pDS->close();
-      return true;
-    }
-
-    if (m_profileManager.GetMasterProfile().getLockMode() != LOCK_MODE_EVERYONE && !g_passwordManager.bMasterUser)
-    {
-      std::map<int, CActor> mapActors;
-
-      while (!m_pDS->eof())
-      {
-        int idActor = m_pDS->fv(0).get_asInt();
-        CActor actor;
-        actor.name = m_pDS->fv(1).get_asString();
-        actor.thumb = m_pDS->fv(2).get_asString();
-        if (idContent != VIDEODB_CONTENT_TVSHOWS)
-        {
-          actor.playcount = m_pDS->fv(3).get_asInt();
-          actor.appearances = 1;
-        }
-        else actor.appearances = m_pDS->fv(4).get_asInt();
-        auto it = mapActors.find(idActor);
-        // is this actor already known?
-        if (it == mapActors.end())
-        {
-          // check path
-          if (g_passwordManager.IsDatabasePathUnlocked(m_pDS->fv("path.strPath").get_asString(),*CMediaSourceSettings::GetInstance().GetSources("video")))
-            mapActors.insert(std::pair<int, CActor>(idActor, actor));
-        }
-        else if (idContent != VIDEODB_CONTENT_TVSHOWS)
-            it->second.appearances++;
-        m_pDS->next();
-      }
-      m_pDS->close();
-
-      for (const auto &i : mapActors)
-      {
-        CFileItemPtr pItem(new CFileItem(i.second.name));
-
-        CVideoDbUrl itemUrl = videoUrl;
-        std::string path = StringUtils::Format("%i/", i.first);
-        itemUrl.AppendPath(path);
-        pItem->SetPath(itemUrl.ToString());
-
-        pItem->m_bIsFolder=true;
-        pItem->GetVideoInfoTag()->SetPlayCount(i.second.playcount);
-        pItem->GetVideoInfoTag()->m_strPictureURL.ParseString(i.second.thumb);
-        pItem->GetVideoInfoTag()->m_iDbId = i.first;
-        pItem->GetVideoInfoTag()->m_type = type;
-        pItem->GetVideoInfoTag()->m_relevance = i.second.appearances;
-        items.Add(pItem);
-      }
-    }
-    else
-    {
-      while (!m_pDS->eof())
-      {
-        try
-        {
-          CFileItemPtr pItem(new CFileItem(m_pDS->fv(1).get_asString()));
-
-          CVideoDbUrl itemUrl = videoUrl;
-          std::string path = StringUtils::Format("%i/", m_pDS->fv(0).get_asInt());
-          itemUrl.AppendPath(path);
-          pItem->SetPath(itemUrl.ToString());
-
-          pItem->m_bIsFolder=true;
-          pItem->GetVideoInfoTag()->m_strPictureURL.ParseString(m_pDS->fv(2).get_asString());
-          pItem->GetVideoInfoTag()->m_iDbId = m_pDS->fv(0).get_asInt();
-          pItem->GetVideoInfoTag()->m_type = type;
-          if (idContent != VIDEODB_CONTENT_TVSHOWS)
-          {
-            // fv(4) is the number of videos watched, fv(3) is the total number.  We set the playcount
-            // only if the number of videos watched is equal to the total number (i.e. every video watched)
-            pItem->GetVideoInfoTag()->SetPlayCount((m_pDS->fv(4).get_asInt() == m_pDS->fv(3).get_asInt()) ? 1 : 0);
-          }
-          pItem->GetVideoInfoTag()->m_relevance = m_pDS->fv(3).get_asInt();
-          if (idContent == VIDEODB_CONTENT_MUSICVIDEOS)
-            pItem->GetVideoInfoTag()->m_artist.emplace_back(pItem->GetLabel());
-          items.Add(pItem);
-          m_pDS->next();
-        }
-        catch (...)
-        {
-          m_pDS->close();
-          CLog::Log(LOGERROR, "%s: out of memory - retrieved %i items", __FUNCTION__, items.Size());
-          return items.Size() > 0;
-        }
-      }
-      m_pDS->close();
-    }
-    CLog::Log(LOGDEBUG, LOGDATABASE, "%s item retrieval took %i ms",
-              __FUNCTION__, XbmcThreads::SystemClockMillis() - time); time = XbmcThreads::SystemClockMillis();
-
-    return true;
-  }
-  catch (...)
-  {
-    m_pDS->close();
-    CLog::Log(LOGERROR, "%s failed", __FUNCTION__);
-  }
-  return false;
-}
-
 bool CVideoDatabase::GetYearsNav(const std::string& strBaseDir, CFileItemList& items, int idContent /* = -1 */, const Filter &filter /* = Filter() */)
 {
   try
@@ -2023,11 +1610,6 @@ std::string CVideoDatabase::GetGenreById(int id)
 std::string CVideoDatabase::GetCountryById(int id)
 {
   return GetSingleValue("country", "name", PrepareSQL("country_id=%i", id));
-}
-
-std::string CVideoDatabase::GetPersonById(int id)
-{
-  return GetSingleValue("actor", "name", PrepareSQL("actor_id=%i", id));
 }
 
 std::string CVideoDatabase::GetStudioById(int id)
